+++ functional/Makefile
@@ -1,6 +1,6 @@
 # Parameters
-WORKLOADS := 200
-TIMEOUT := 190
+WORKLOADS := 50
+TIMEOUT := 60
 TTL := 60
 
 # Convenience vars
@@ -25,11 +25,12 @@ agent:
 	cd .. && touch $(AGENT_LOG)
 	$(eval TOKEN := $(shell docker-compose exec runner $(DOCKER_SPIRE)/functional/tools/tokengenerator/tokengenerator))
 	@echo Generated join token is $(TOKEN)
+	docker-compose exec runner bash -c 'cd $(DOCKER_SPIRE)/functional && tools/stresstest/stresstest createusers -token=$(TOKEN) -workloads $(WORKLOADS) -ttl $(TTL)'
 	docker-compose exec -d runner bash -c 'cd $(DOCKER_SPIRE) && cmd/spire-agent/spire-agent run -joinToken $(TOKEN) -logLevel DEBUG -logFile /spire-orig/$(AGENT_LOG)'
 	sleep 5
 
 test:
-	docker-compose exec runner bash -c 'cd $(DOCKER_SPIRE)/functional && tools/stresstest/stresstest -token=$(TOKEN) -workloads $(WORKLOADS) -timeout $(TIMEOUT) -ttl $(TTL)'
+	docker-compose exec runner bash -c 'cd $(DOCKER_SPIRE)/functional && tools/stresstest/stresstest run -token=$(TOKEN) -workloads $(WORKLOADS) -timeout $(TIMEOUT) -ttl $(TTL)'
 	docker-compose exec runner bash -c 'cp $(DOCKER_SPIRE)/functional/*.log /spire-orig/functional/'
 	docker-compose down
